feat: export applyResolved from the viewer
